requirements.txt
pytest >= 7.4.0
numpy >=2.2
pillow == 12.0.0